Removed frontend package and fixed window
Having a whole frontend package with structs and traits to implement was proving to be a major pain to integrate with other libraries.

With the switch to piston, draw and update logic are handled in main.rs. In the future, a separate crate for the emulator core (library) and application (frontend) could be better.